// src/configs/g_types/types.ts
export type NavSectionTitle = {
  action?: string;
  subject?: string;
  sectionTitle: string;
};

export type NavGroup = {
  icon?: string;
  title: string;
  action?: string;
  subject?: string;
  badgeContent?: string;
  children?: (NavGroup | NavLink)[];
  badgeColor?:
  | "default"
  | "primary"
  | "secondary"
  | "success"
  | "error"
  | "warning"
  | "info";
};

export type NavLink = {
  icon?: string;
  path?: string;
  title: string;
  action?: string;
  subject?: string;
  disabled?: boolean;
  badgeContent?: string;
  externalLink?: boolean;
  openInNewTab?: boolean;
  badgeColor?:
  | "default"
  | "primary"
  | "secondary"
  | "success"
  | "error"
  | "warning"
  | "info";
};

export type VerticalNavItemsType = (NavLink | NavGroup | NavSectionTitle)[];